--- a/test/regressions/test_upsample_bilinear_bwd.py
+++ b/test/regressions/test_upsample_bilinear_bwd.py
@@ -1,6 +1,5 @@
 # Owner(s): ["module: intel"]
 import torch
-from torch.autograd import Variable
 from torch.testing._internal.common_utils import TestCase
 
 device = torch.device("xpu")
@@ -17,12 +16,23 @@ class TestTorchMethod(TestCase):
             )
             for memory_format in [torch.channels_last, torch.contiguous_format]:
                 for align_corners in [True, False]:
-                    grad_output_cpu = grad_output_cpu.contiguous(memory_format=memory_format)
-                    r_cpu = torch._ops.ops.aten.upsample_bilinear2d_backward(grad_output_cpu.to(torch.float64), output_size=(512, 512),input_size=(1,3,256,256),align_corners=align_corners)
-                    r_xpu = torch._ops.ops.aten.upsample_bilinear2d_backward(grad_output_cpu.to("xpu"), output_size=(512, 512),input_size=(1,3,256,256),align_corners=align_corners)
-                    
-                    self.assertEqual(r_cpu.to(dtype), r_xpu.cpu())
+                    grad_output_cpu = grad_output_cpu.contiguous(
+                        memory_format=memory_format
+                    )
+                    r_cpu = torch._ops.ops.aten.upsample_bilinear2d_backward(
+                        grad_output_cpu.to(torch.float64),
+                        output_size=(512, 512),
+                        input_size=(1, 3, 256, 256),
+                        align_corners=align_corners,
+                    )
+                    r_xpu = torch._ops.ops.aten.upsample_bilinear2d_backward(
+                        grad_output_cpu.to("xpu"),
+                        output_size=(512, 512),
+                        input_size=(1, 3, 256, 256),
+                        align_corners=align_corners,
+                    )
 
+                    self.assertEqual(r_cpu.to(dtype), r_xpu.cpu())
 
         for dtype in test_dtypes:
             _test_upsample_bilinear_bwd(dtype)