We do not need to use atomic add for this op. Currently I only implement the case of align corners = True

Before: These kernels take 31ms out of 82ms in torchbench pytorch_unet fp16 training.
Now they takes 2.26ms out of 51ms....